Move step functionality to cli-engine
