+++ packages/cli/libs/commands/run/step.command.spec.ts
@@ -1,8 +1,10 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import { step } from "./step.command";
-import { printStepResult } from "../../print.service";
-import { of } from "rxjs";
 
+jest.mock("../../data-access/workflow.dao", () => ({
+  readWorkflowOrTerminate: () => Promise.resolve("my-workflow"),
+  readWorkflowNamesOrTerminate: () => Promise.resolve(["my-workflow"]),
+}));
 jest.mock("../../data-access/config.dao");
 jest.mock("../../data-access/state.dao", () => ({
   readStateOrTerminate: () => Promise.resolve({}),
@@ -14,70 +16,24 @@ jest.mock("../../data-access/environment.dao", () => ({
 jest.mock("../../validator", () => ({
   validateEnvironmentOrTerminate: () => Promise.resolve(),
 }));
-jest.mock("../../data-access/workflow.dao");
-jest.mock("../../inquirer.service");
-jest.mock("@jyfti/engine", () => {
-  const engine = {
-    isComplete: jest.fn(() => true),
-    step: jest.fn(() => require("rxjs").empty()),
-    transitionFrom: jest.fn(() => (stepResult$) =>
-      stepResult$.pipe(require("rxjs/operators").map(() => ({})))
-    ),
-    resolveStep: () => ({}),
-  };
-  return {
-    engine,
-    createEngine: () => engine,
-    isSuccess: jest.requireActual("@jyfti/engine").isSuccess,
-    isFailure: jest.requireActual("@jyfti/engine").isFailure,
-  };
-});
+jest.mock("../../inquirer.service", () => ({
+  promptWorkflow: jest.fn(() => Promise.resolve("my-workflow")),
+}));
+jest.mock("../../cli-engine", () => ({
+  runStep: jest.fn(() => Promise.resolve()),
+}));
 
 describe("the step command", () => {
-  let logSpy: any;
-  let errorSpy: any;
-  let writeStateSpy: any;
-
-  beforeEach(() => {
-    logSpy = jest.spyOn(console, "log").mockImplementation(() => {});
-    errorSpy = jest.spyOn(console, "error").mockImplementation(() => {});
-    writeStateSpy = require("../../data-access/state.dao").writeState;
-  });
-
-  it("should execute the next step", async () => {
-    require("@jyfti/engine").engine.isComplete.mockReturnValue(false);
-    require("@jyfti/engine").engine.step.mockReturnValue(
-      of({
-        path: [0],
-        evaluation: "a",
-      })
+  it("should prompt for workflow name and continue", async () => {
+    const runStep = require("../../cli-engine").runStep;
+    runStep.mockReturnValue(Promise.resolve());
+    await step(undefined);
+    expect(
+      require("../../inquirer.service").promptWorkflow
+    ).toHaveBeenCalledWith(
+      ["my-workflow"],
+      "Which workflow do you want to progress?"
     );
-    await step("my-workflow", { verbose: false });
-    expect(logSpy).toHaveBeenCalledWith(
-      printStepResult({ path: [0], evaluation: null })
-    );
-    expect(errorSpy).toHaveBeenCalledTimes(0);
-    expect(writeStateSpy).toHaveBeenCalledTimes(1);
-  });
-
-  it("should return if already completed", async () => {
-    require("@jyfti/engine").engine.isComplete.mockReturnValue(true);
-    await step("my-workflow", { verbose: false });
-    expect(logSpy).toHaveBeenCalledWith("Workflow execution already completed");
-    expect(errorSpy).toHaveBeenCalledTimes(0);
-    expect(writeStateSpy).toHaveBeenCalledTimes(0);
-  });
-
-  it("should log a failed state if the engine returns an error", async () => {
-    const stepResult = {
-      path: [0],
-      error: new Error("Something went wrong."),
-    };
-    require("@jyfti/engine").engine.isComplete.mockReturnValue(false);
-    require("@jyfti/engine").engine.step.mockReturnValue(of(stepResult));
-    await step("my-workflow", { verbose: false });
-    expect(logSpy).toHaveBeenCalledWith(printStepResult(stepResult));
-    expect(errorSpy).toHaveBeenCalledTimes(0);
-    expect(writeStateSpy).toHaveBeenCalledTimes(0);
+    expect(runStep).toHaveBeenCalledTimes(1);
   });
 });
